Re-organize javascript files structure
